feat: move useValidateCredential to common
