+++ packages/common/package.json
@@ -41,6 +41,7 @@
     "swr": "^2.0.1",
     "tailwind-styled-components": "^2.2.0",
     "tiny-emitter": "^2.1.0",
+    "ts-debounce": "^4.0.0",
     "typescript": "latest",
     "viem": "^1.5.3",
     "wagmi": "0.12.19",
