+++ packages/grant-explorer/src/features/round/ViewProjectDetails.tsx
@@ -1,14 +1,10 @@
 import { datadogLogs } from "@datadog/browser-logs";
-import {
-  PROVIDER_ID,
-  VerifiableCredential,
-} from "@gitcoinco/passport-sdk-types";
 import { ShieldCheckIcon } from "@heroicons/react/24/solid";
 import {
-  PassportVerifierWithExpiration,
   formatDateWithOrdinal,
   renderToHTML,
   useParams,
+  useValidateCredential,
 } from "common";
 import { getAlloVersion } from "common/src/config";
 import { formatDistanceToNowStrict } from "date-fns";
@@ -22,7 +18,6 @@ import React, {
   useMemo,
   useState,
 } from "react";
-import useSWR from "swr";
 import { useAccount, useEnsName } from "wagmi";
 import DefaultLogoImage from "../../assets/default_logo.png";
 import { ReactComponent as GithubIcon } from "../../assets/github-logo.svg";
@@ -42,7 +37,7 @@ import { useGap } from "../api/gap";
 import { StatList } from "./OSO/ImpactStats";
 import { useOSO } from "../api/oso";
 import { CheckIcon, ShoppingCartIcon } from "@heroicons/react/24/outline";
-import { DataLayer, useDataLayer } from "data-layer";
+import { useDataLayer } from "data-layer";
 import { DefaultLayout } from "../common/DefaultLayout";
 import { truncate } from "../common/utils/truncate";
 import {
@@ -71,15 +66,6 @@ const CalendarIcon = (props: React.SVGProps<SVGSVGElement>) => {
     </svg>
   );
 };
-
-enum VerifiedCredentialState {
-  VALID,
-  INVALID,
-  PENDING,
-}
-
-export const IAM_SERVER =
-  "did:key:z6MkghvGHLobLEdj1bgRLhS4LPGJAvbMA1tn2zcRyqmYU5LC";
 
 const useProjectDetailsParams = useParams<{
   chainId: string;
@@ -323,36 +309,6 @@ function ProjectDetailsTabs(props: {
   );
 }
 
-function useVerifyProject(project?: Project) {
-  const { credentials = {} } = project?.projectMetadata ?? {};
-  const dataLayer = useDataLayer();
-
-  // Return data as { twitter?: boolean, ... }
-  return useSWR<{ [K in Lowercase<PROVIDER_ID>]?: boolean }>(credentials, () =>
-    Promise.all(
-      // Check verifications for all credentials in project metadata
-      Object.entries(credentials).map(async ([provider, credential]) => ({
-        provider,
-        verified: await isVerified({
-          dataLayer,
-          verifiableCredential: credential,
-          provider,
-          project,
-        }),
-      }))
-    ).then((verifications) =>
-      // Convert to object ({ [provider]: isVerified })
-      verifications.reduce(
-        (acc, x) => ({
-          ...acc,
-          [x.provider]: x.verified === VerifiedCredentialState.VALID,
-        }),
-        {}
-      )
-    )
-  );
-}
-
 function ProjectLinks({ project }: { project?: Project }) {
   const {
     recipient,
@@ -362,13 +318,22 @@ function ProjectLinks({ project }: { project?: Project }) {
       projectTwitter,
       projectGithub,
       userGithub,
+      credentials,
     },
   } = project ?? { projectMetadata: {} };
 
   // @ts-expect-error Temp until viem (could also cast recipient as Address or update the type)
   const ens = useEnsName({ address: recipient, enabled: Boolean(recipient) });
 
-  const verified = useVerifyProject(project);
+  const { isValid: validTwitterCredential } = useValidateCredential(
+    credentials?.twitter,
+    projectTwitter
+  );
+
+  const { isValid: validGithubCredential } = useValidateCredential(
+    credentials?.github,
+    projectGithub
+  );
 
   const createdOn =
     createdAt &&
@@ -392,7 +357,7 @@ function ProjectLinks({ project }: { project?: Project }) {
         <ProjectLink
           url={`https://twitter.com/${projectTwitter}`}
           icon={TwitterIcon}
-          isVerified={verified.data?.twitter}
+          isVerified={validTwitterCredential}
         >
           {projectTwitter}
         </ProjectLink>
@@ -401,7 +366,7 @@ function ProjectLinks({ project }: { project?: Project }) {
         <ProjectLink
           url={`https://github.com/${projectGithub}`}
           icon={GithubIcon}
-          isVerified={verified.data?.github}
+          isVerified={validGithubCredential}
         >
           {projectGithub}
         </ProjectLink>
@@ -642,62 +607,4 @@ function CartButtonToggle(props: {
       {props.isAlreadyInCart ? "Added to cart" : "Add to cart"}
     </button>
   );
-}
-
-function vcProviderMatchesProject(
-  provider: string,
-  verifiableCredential: VerifiableCredential,
-  project: Project | undefined
-) {
-  let vcProviderMatchesProject = false;
-  if (provider === "twitter") {
-    vcProviderMatchesProject =
-      verifiableCredential.credentialSubject.provider
-        ?.split("#")[1]
-        .toLowerCase() ===
-      project?.projectMetadata.projectTwitter?.toLowerCase();
-  } else if (provider === "github") {
-    vcProviderMatchesProject =
-      verifiableCredential.credentialSubject.provider
-        ?.split("#")[1]
-        .toLowerCase() ===
-      project?.projectMetadata.projectGithub?.toLowerCase();
-  }
-  return vcProviderMatchesProject;
-}
-
-function vcIssuedToAddress(vc: VerifiableCredential, address: string) {
-  const vcIdSplit = vc.credentialSubject.id.split(":");
-  const addressFromId = vcIdSplit[vcIdSplit.length - 1];
-  return addressFromId === address;
-}
-
-async function isVerified(args: {
-  verifiableCredential: VerifiableCredential;
-  provider: string;
-  project: Project | undefined;
-  dataLayer: DataLayer;
-}) {
-  const { verifiableCredential, provider, project } = args;
-
-  const passportVerifier = new PassportVerifierWithExpiration();
-  const vcHasValidProof =
-    await passportVerifier.verifyCredential(verifiableCredential);
-
-  const vcIssuedByValidIAMServer = verifiableCredential.issuer === IAM_SERVER;
-  const providerMatchesProject = vcProviderMatchesProject(
-    provider,
-    verifiableCredential,
-    project
-  );
-  const vcIssuedToAtLeastOneProjectOwner = (
-    project?.projectMetadata?.owners ?? []
-  ).some((owner) => vcIssuedToAddress(verifiableCredential, owner.address));
-
-  return vcHasValidProof &&
-    vcIssuedByValidIAMServer &&
-    providerMatchesProject &&
-    vcIssuedToAtLeastOneProjectOwner
-    ? VerifiedCredentialState.VALID
-    : VerifiedCredentialState.INVALID;
 }
